[MCUX-87599] Fix Coverity and MISRA issues in conn, hci_core, smp, spp, and shell
- bt_pal_conn.c: Add explicit casting to ensure MIN() results are
  correctly typed
- bt_pal_hci_core.c: Fix incorrect bitmask operation on connection handle
- bt_pal_smp.c: Remove redundant conditional block that triggers warnings
- bt_pal_spp.c: Unify error handling logic for RFCOMM control requests;
  ensure consistent use of return values and logging
- fsl_shell.c: Fix missing brace logic in auto-complete path; simplify
  and correct parameter length handling to avoid buffer issues flagged
  by static analysis

+++ source/impl/ethermind/host/bt_pal_conn.c
@@ -678,7 +678,7 @@ static int send_buf(struct bt_conn *conn, struct net_buf *buf,
 	tx->cb = (bt_conn_tx_cb_t)cb;
 	tx->user_data = ud;
 
-	uint16_t frag_len = MIN(conn_mtu(conn), len);
+	uint16_t frag_len = (uint16_t)MIN(conn_mtu(conn), len);
 
 	__ASSERT_NO_MSG(buf->ref == 1);
